Cleanup code to reduce number of queries
