+++ app/javascript/toby/views/resource/index.js
@@ -1,34 +1,19 @@
-import * as Sentry from "@sentry/react";
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import queryString from "query-string";
-import { Box, Text } from "@advisable/donut";
-import { Exclamation } from "@styled-icons/heroicons-solid/Exclamation";
+import { Box } from "@advisable/donut";
 import { Adjustments } from "@styled-icons/heroicons-solid/Adjustments";
 import { motion } from "framer-motion";
-import { useHistory, useLocation } from "react-router-dom";
-import { useBottomScrollListener } from "react-bottom-scroll-listener";
-import { useFetchResources } from "../../utilities";
-import {
-  StyledLayout,
-  StyledHeader,
-  StyledHeaderRow,
-  StyledHeaderCell,
-  StyledRow,
-  StyledCell,
-  StyledViewport,
-  StyledScrollContainer,
-} from "../../styles";
+import { useLocation } from "react-router-dom";
+import { StyledLayout, StyledHeader, StyledViewport } from "../../styles";
 import FilterDrawer from "./Filters";
 import Navigation from "../../components/Navigation";
-import { Attribute } from "../../attributes";
 import DetailsModal from "./DetailsModal";
 import LoadingIndicator from "src/components/Loading";
-import { useNotifications } from "src/components/Notifications";
-import Loading from "./Loading";
 import { useResourceViews, useUpdateView } from "../../queries";
 import HeaderButton from "../../components/HeaderButton";
 import ViewsDropdown from "./ViewsDropdown";
 import SortMenu from "./SortMenu";
+import Records from "./Records";
 
 export default function ResourceConfig({ resource }) {
   const { loading, data } = useResourceViews(resource.type);
@@ -37,23 +22,7 @@ export default function ResourceConfig({ resource }) {
   return <Resource resource={resource} views={data.views} />;
 }
 
-function APIError() {
-  return (
-    <Box
-      margin={4}
-      padding={4}
-      maxWidth="400px"
-      color="neutral800"
-      display="inline-flex"
-      alignItems="center"
-    >
-      <Exclamation size={20} />
-      <Text ml={2}>Failed to fetch records</Text>
-    </Box>
-  );
-}
-
-function initializeViewFilters(view) {
+function initFilters(view) {
   if (!view) return [];
 
   return (
@@ -66,9 +35,7 @@ function initializeViewFilters(view) {
 }
 
 function Resource({ resource, views }) {
-  const history = useHistory();
   const location = useLocation();
-  const { error: notifyError } = useNotifications();
   const [updateView] = useUpdateView();
 
   const currentView = useMemo(() => {
@@ -76,22 +43,12 @@ function Resource({ resource, views }) {
     if (!queryParams.view) return null;
     return views.find((v) => v.id === queryParams.view);
   }, [views, location.search]);
+
   const [sortBy, setSortBy] = useState(currentView?.sortBy || "created_at");
   const [sortOrder, setSortOrder] = useState(currentView?.sortOrder || "ASC");
-  const [filters, setFilters] = useState(() =>
-    initializeViewFilters(currentView),
-  );
+  const [filters, setFilters] = useState(() => initFilters(currentView));
 
   const [isOpen, setIsOpen] = useState(false);
-  const { loading, data, fetchMore, error } = useFetchResources(
-    resource,
-    filters,
-    sortBy,
-    sortOrder,
-  );
-
-  const hasNextPage = data?.records.pageInfo.hasNextPage;
-  const endCursor = data?.records.pageInfo.endCursor;
 
   useEffect(() => {
     const queryParams = queryString.parse(location.search);
@@ -100,14 +57,8 @@ function Resource({ resource, views }) {
       : null;
     setSortBy(currentView?.sortBy || "created_at");
     setSortOrder(currentView?.sortOrder || "ASC");
-    setFilters(initializeViewFilters(currentView));
+    setFilters(initFilters(currentView));
   }, [location.search]);
-
-  useEffect(() => {
-    if (error) {
-      notifyError("Failed to load records");
-    }
-  }, [error, notifyError]);
 
   const handleUpdateView = useCallback(
     (changes) => {
@@ -133,27 +84,19 @@ function Resource({ resource, views }) {
     [handleUpdateView],
   );
 
-  const scrollRef = useBottomScrollListener(() => {
-    if (!loading && !hasNextPage) return;
-    fetchMore({ variables: { cursor: endCursor } });
-  });
-
-  const edges = data?.records.edges || [];
-
-  const openRecord = (id) => () => {
-    history.push({
-      ...location,
-      pathname: `${location.pathname}/${id}`,
-    });
-  };
-
   return (
     <StyledLayout>
       <DetailsModal resource={resource} />
       <StyledHeader>
         <Navigation resource={resource} />
         <Box display="flex" alignItems="center" paddingLeft="8px">
-          <ViewsDropdown views={views} resource={resource} filters={filters} />
+          <ViewsDropdown
+            views={views}
+            resource={resource}
+            filters={filters}
+            sortBy={sortBy}
+            sortOrder={sortOrder}
+          />
           <HeaderButton icon={Adjustments} onClick={() => setIsOpen(!isOpen)}>
             Filters
             {filters.length > 0 && (
@@ -174,10 +117,10 @@ function Resource({ resource, views }) {
           </HeaderButton>
           <SortMenu
             sortBy={sortBy}
+            resource={resource}
             setSortBy={setSortBy}
             sortOrder={sortOrder}
             setSortOrder={setSortOrder}
-            resource={resource}
             handleUpdateView={handleUpdateView}
           />
         </Box>
@@ -185,54 +128,23 @@ function Resource({ resource, views }) {
       <StyledViewport>
         <FilterDrawer
           views={views}
-          resource={resource}
           open={isOpen}
           filters={filters}
+          resource={resource}
           onApply={handleUpdateFilters}
         />
-        <StyledScrollContainer
-          ref={scrollRef}
-          as={motion.div}
-          $filterOpen={isOpen}
+        <motion.div
           transition={{ duration: 0.2 }}
           animate={{ x: isOpen ? 400 : 0 }}
+          style={{ width: isOpen ? "calc(100vw - 400px)" : "100vw" }}
         >
-          {/* Dear future developer. I know this inline-block looks random. But its important. */}
-          <Box display="inline-block" minWidth="100vw">
-            <StyledHeaderRow>
-              {resource.attributes.map((attr) => (
-                <StyledHeaderCell key={attr.name}>
-                  {attr.columnLabel}
-                </StyledHeaderCell>
-              ))}
-            </StyledHeaderRow>
-            <Box>
-              {error && <APIError>The API returned an error</APIError>}
-              {!error &&
-                edges.map(({ node }) => (
-                  <StyledRow key={node.id} onClick={openRecord(node.id)}>
-                    {resource.attributes.map((attr) => (
-                      <StyledCell key={attr.name}>
-                        <Sentry.ErrorBoundary
-                          fallback={
-                            <Box display="inline-flex" alignItems="center">
-                              <Exclamation size={16} />
-                              <Text mt="-1px" ml={1}>
-                                Error
-                              </Text>
-                            </Box>
-                          }
-                        >
-                          <Attribute record={node} attribute={attr} />
-                        </Sentry.ErrorBoundary>
-                      </StyledCell>
-                    ))}
-                  </StyledRow>
-                ))}
-              {loading && <Loading resource={resource} />}
-            </Box>
-          </Box>
-        </StyledScrollContainer>
+          <Records
+            resource={resource}
+            filters={filters}
+            sortBy={sortBy}
+            sortOrder={sortOrder}
+          />
+        </motion.div>
       </StyledViewport>
     </StyledLayout>
   );
